Added Merging pck into exe More help Fixed sources link in about Added limit to backgroundProcess Better folder scan function Added PCKVersion class More error checks PCK offsets updating simplified Split now use not "new pair name" but full path to exe Extract and pack now can operate only selected files too Rip and Split now can delete and restore backup and output files

+++ Tests/LockedFile.cs
@@ -6,17 +6,21 @@ namespace Tests
     internal class LockedFile : IDisposable
     {
         private FileStream file;
+        bool DeleteAfterDispose = true;
 
-        public LockedFile(string path)
+        public LockedFile(string path, bool deleteAfterDispose = true)
         {
             file = File.OpenWrite(path);
+            DeleteAfterDispose = deleteAfterDispose;
         }
 
         public void Dispose()
         {
             var path = file.Name;
             file.Close();
-            File.Delete(path);
+
+            if (DeleteAfterDispose)
+                File.Delete(path);
         }
     }
 }
